+++ cacti/nuca.cc
@@ -204,7 +204,7 @@ Nuca::sim_nuca()
     }
   }
 
-  iterations = (int)logtwo((int)g_ip->cache_sz/MIN_BANKSIZE);
+  iterations = (int)logtwo((int)(g_ip->cache_sz/MIN_BANKSIZE));
 
   if (g_ip->force_wiretype)
   {
@@ -362,15 +362,15 @@ Nuca::sim_nuca()
 
         if (it < 7) {
           nuca_list.back()->nuca_pda.delay = opt_acclat +
-            cont_stats[l2_c][core_in][ro][it][num_cyc/2-1];
+            cont_stats[l2_c][core_in][ro][it][(int)(num_cyc/2-1)];
           nuca_list.back()->contention =
-            cont_stats[l2_c][core_in][ro][it][num_cyc/2-1];
+            cont_stats[l2_c][core_in][ro][it][(int)(num_cyc/2-1)];
         }
         else {
           nuca_list.back()->nuca_pda.delay = opt_acclat +
-            cont_stats[l2_c][core_in][ro][7][num_cyc/2-1];
+            cont_stats[l2_c][core_in][ro][7][(int)(num_cyc/2-1)];
           nuca_list.back()->contention =
-            cont_stats[l2_c][core_in][ro][7][num_cyc/2-1];
+            cont_stats[l2_c][core_in][ro][7][(int)(num_cyc/2-1)];
         }
         nuca_list.back()->nuca_pda.power.readOp.dynamic = opt_dyn_power;
         nuca_list.back()->nuca_pda.power.readOp.leakage = opt_leakage_power;
